Move out swap details to separate hook
