+++ src/hooks/index.js
@@ -34,6 +34,7 @@ export { default as useSendableUniqueTokens } from './useSendableUniqueTokens';
 export { default as useSendSavingsAccount } from './useSendSavingsAccount';
 export { default as useShakeAnimation } from './useShakeAnimation';
 export { default as useShowcaseTokens } from './useShowcaseTokens';
+export { default as useSwapDetails } from './useSwapDetails';
 export { default as useTimeout } from './useTimeout';
 export { default as useTopMovers } from './useTopMovers';
 export { default as useTransactionConfirmation } from './useTransactionConfirmation';
